+++ packages/client/src/actions/projects.ts
@@ -4,18 +4,17 @@ import {
   getRetroFundingRoundProjects,
   getRetroFundingRoundProjectsResponse,
   getRetroFundingRoundProjectById,
-  getRetroFundingRoundProjectByIdResponse
+  getRetroFundingRoundProjectByIdResponse,
 } from "@/__generated__/api/agora";
 import {
   Project as OpProject,
-  PageMetadata
+  PageMetadata,
 } from "@/__generated__/api/agora.schemas";
 
 import { parseOpProjectToProjectItem } from "@/utils/parseData";
 
 import { getProjectEndorsements } from "./endorsements";
 import { getOsoCodeMetricsByArtifact } from "./repos";
-import { Project } from "@/types/projects";
 
 const getArtifactNameAndNamespace = (url: string) => {
   const urlParts = url.split("/");
@@ -24,14 +23,14 @@ const getArtifactNameAndNamespace = (url: string) => {
 
   return {
     artifactName,
-    artifactNamespace
+    artifactNamespace,
   };
 };
 
 export type ProjectsResponse = {
   metadata?: PageMetadata;
   data?: OpProject[];
-}
+};
 
 export type ProjectResponse = {
   data?: OpProject;
@@ -39,12 +38,12 @@ export type ProjectResponse = {
 
 export const getProjects = async (
   query?: string,
-  page?: number
+  page?: number,
 ): Promise<Project[]> => {
   const projects = await getRetroFundingRoundProjects(5, {
     limit: 25,
     offset: page ? (page - 1) * 25 : 0,
-    category: "all"
+    category: "all",
   }).then((results: getRetroFundingRoundProjectsResponse) => {
     const res: ProjectsResponse = results.data;
     return res.data;
@@ -57,26 +56,26 @@ export const getProjects = async (
 };
 
 export const getProject = async (
-  projectId: string
+  projectId: string,
 ): Promise<Project | undefined> => {
   const opProject = await getRetroFundingRoundProjectById(5, projectId).then(
     (results: getRetroFundingRoundProjectByIdResponse) => {
       return results.data;
-    }
+    },
   );
 
   if (!opProject) return undefined;
 
   const project = parseOpProjectToProjectItem(opProject);
   const repos = project.repositories?.map((repo: string) =>
-    getArtifactNameAndNamespace(repo)
+    getArtifactNameAndNamespace(repo),
   );
 
   const endorsements = await getProjectEndorsements(projectId);
   const metrics = await getOsoCodeMetricsByArtifact(
     projectId,
     repos[0]?.artifactName,
-    repos[0]?.artifactNamespace
+    repos[0]?.artifactNamespace,
   );
 
   return {
