+++ packages/client/src/components/Project/ListItem.tsx
@@ -1,10 +1,9 @@
 "use client";
 import React from "react";
 import Image from "next/image";
+import { useRouter } from "next/navigation";
 
 import { formatLastUpdated } from "@/utils/text";
-import { useRouter } from "next/navigation";
-import { ProjectCategory } from "@/types/projects";
 
 export interface ProjectListItemProps {
   id: string;
