+++ packages/client/src/types/app.d.ts
@@ -1,5 +1,3 @@
-import { type introspection_types } from "./oso-env";
-
 declare interface User {
   fid?: number;
   name?: string;
@@ -47,7 +45,70 @@ declare interface CreateMetric {
   keyword: string;
 }
 
-interface CreateEndorsement {
+declare type GrantTrack =
+  | "onchain-builders"
+  | "op-stack"
+  | "dev-tooling"
+  | "governance";
+
+declare type ProjectCategory =
+  | "cefi"
+  | "cross-chain"
+  | "defi"
+  | "governance"
+  | "nft"
+  | "social"
+  | "utility";
+
+declare type Contract = {
+  id: string;
+  contractAddress: string;
+  deployerAddress: string;
+  deploymentHash: string;
+  chainId: number;
+};
+
+declare interface ProjectGrant {
+  title: string;
+  date: string;
+  funds_received: string;
+  description: string;
+  link?: string;
+}
+
+// TODO hacked with key:string until we have data fields narrowed down
+declare interface ProjectItem {
+  id: string;
+  title: string;
+  avatar_image: string;
+  category: ProjectCategory;
+  updated_at: string;
+  creator: string;
+  [key: string]: any;
+}
+
+declare interface Project extends ProjectItem {
+  grant_track: GrantTrack;
+  description: string;
+  banner_image: string;
+  metrics: ProjectMetric[];
+  endorsements: Endorsement[];
+  repositories: string[];
+  contracts: Contract[];
+  funding: ProjectGrant[];
+  socials: string[];
+}
+
+declare enum ProjectMetricSource {
+  OSO = "Open Source Observer",
+  EAS = "EAS",
+}
+
+declare interface ProjectMetric {
+  source: ProjectMetricSource;
+}
+
+declare interface CreateEndorsement {
   recipient?: string; // Contract for Project
   projectUID: string;
   metricUID?: string;
