+++ packages/client/src/views/Projects/Details.tsx
@@ -15,8 +15,6 @@ import { ProjectOverview } from "@/components/Project/Overview";
 import { ProjectAttestations } from "@/components/Project/Attestations";
 import { ProjectEndorsements } from "@/components/Project/Endorsements";
 import { AttestFormValues } from "@/components/Project/Attest/Metric";
-import { Project } from "@/types/projects";
-import { User } from "@/types/app";
 
 export interface ProjectViewProps {
   user: User;
